git-svn-id: http://vps6982.inmotionhosting.com/svn/lasi@743 49725315-0998-4498-bb4a-373bc1a52cb2

+++ LASI_Algorithm/LexicalStructures/NounRelatedConstructs/NPAggregateSubjectObject.cs
@@ -16,128 +16,106 @@ namespace LASI.Algorithm
     /// <seealso cref="IEntity"/>
     public class EntityGroup : IEntityGroup
     {
-        public EntityGroup(IEnumerable<IEntity> aggregated)
-        {
+        public EntityGroup(IEnumerable<IEntity> aggregated) {
             AggregatesEntities = aggregated;
         }
-        public void AddPossession(IEntity possession)
-        {
+        public void AddPossession(IEntity possession) {
             if (!_possessions.Contains(possession)) {
                 _possessions.Add(possession);
             }
         }
 
         /// <summary>
-        /// Gets or sets the Entity NounPointerSymbol; Person, Place, Thing, Organization, or Activity; of the EntityGroup instance.
+        /// Gets or sets the Entity Noun; Person, Place, Thing, Organization, or Activity; of the EntityGroup instance.
         /// </summary>
-        public EntityKind EntityKind
-        {
+        public EntityKind EntityKind {
             get;
             set;
         }
         /// <summary>
         /// Gets the IVerbal instance, generally a TransitiveVerb or TransitiveVerbPhrase, which the EntityGroup is the DIRECT object of.
         /// </summary>
-        public IVerbal DirectObjectOf
-        {
+        public IVerbal DirectObjectOf {
             get;
             set;
         }
         /// <summary>
         /// Gets the IVerbal instance, generally a TransitiveVerb or TransitiveVerbPhrase, which the EntityGroup is the INDIRECT object of.
         /// </summary>
-        public IVerbal IndirectObjectOf
-        {
+        public IVerbal IndirectObjectOf {
             get;
             set;
         }
         /// <summary>
         /// Gets the IVerbal instance, generally a Verb or VerbPhrase, which the EntityGroup is the subject of.
         /// </summary>
-        public IVerbal SubjectOf
-        {
+        public IVerbal SubjectOf {
             get;
             set;
         }
 
-        public void BindPronoun(IPronoun pro)
-        {
+        public void BindPronoun(IPronoun pro) {
             if (!BoundPronouns.Contains(pro)) {
                 _boundPronouns.Add(pro);
             }
         }
 
-        public IEnumerable<IPronoun> BoundPronouns
-        {
-            get
-            {
+        public IEnumerable<IPronoun> BoundPronouns {
+            get {
                 return _boundPronouns;
             }
         }
 
-        public void BindDescriptor(IDescriptor adj)
-        {
+        public void BindDescriptor(IDescriptor adj) {
             if (!_describers.Contains(adj)) {
                 _describers.Add(adj);
             }
         }
 
-        public IEnumerable<IDescriptor> Descriptors
-        {
-            get
-            {
+        public IEnumerable<IDescriptor> Descriptors {
+            get {
                 return _describers;
             }
         }
         /// <summary>
         /// Gets all of the constructs the NPAggregate can be determined to "own" collectively 
         /// /// </summary>
-        public IEnumerable<IEntity> Possessed
-        {
-            get
-            {
+        public IEnumerable<IEntity> Possessed {
+            get {
                 return _possessions;
             }
         }
 
 
-        public override string ToString()
-        {
+        public override string ToString() {
             return base.ToString() + String.Format(" \"{0}\"", Text);
         }
-        public IEntity Possesser
-        {
+        public IEntity Possesser {
             get;
             set;
         }
 
-        public string Text
-        {
-            get
-            {
+        public string Text {
+            get {
                 return AggregatesEntities.Aggregate("", (aggr, ent) => aggr += ent + " ").Trim();
             }
         }
-        public Type Type
-        {
-            get
-            {
+        public Type Type {
+            get {
                 return GetType();
             }
         }
         /// <summary>
         /// Gets or sets the numeric Weight of the Noun-Phrase-Aggregate within the context of its document.
         /// </summary>
-        public decimal Weight
-        {
+        public double Weight {
             get;
             set;
         }
         /// <summary>
         /// Gets or sets the numeric Weight of the Noun-Phrase-Aggregate over the context of all extant documents.
         /// </summary>
-        public decimal MetaWeight
-        {
+        public double MetaWeight {
             get;
             set;
         }
@@ -147,19 +125,16 @@ namespace LASI.Algorithm
         /// <summary>
         /// Gets the EnumerableCollection of Entities compose to form the aggregate
         /// </summary>
-        public IEnumerable<IEntity> AggregatesEntities
-        {
+        public IEnumerable<IEntity> AggregatesEntities {
             get;
             protected set;
         }
 
-        public IEnumerator<IEntity> GetEnumerator()
-        {
+        public IEnumerator<IEntity> GetEnumerator() {
             return AggregatesEntities.GetEnumerator();
         }
 
-        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator()
-        {
+        System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() {
             throw new NotImplementedException();
         }
